Finished code that outputs the correct info.

+++ src/thoughtworkschallenge/ParsedItems.java
@@ -3,11 +3,11 @@ package thoughtworkschallenge;
 import java.math.BigDecimal;
 
 public class ParsedItems {
-    final Float numberBought;
+    final int numberBought;
     final String itemBought;
     final BigDecimal price;
 
-    public ParsedItems(Float numberBought, String itemBought, BigDecimal price) {
+    public ParsedItems(int numberBought, String itemBought, BigDecimal price) {
         this.numberBought = numberBought;
         this.itemBought = itemBought;
         this.price = price;
